requirements.txt
pypdf
scikit-learn
pandas 
numpy 
pyresparser
tensorflow
nltk
streamlit 
sentence-transformer
rapidfuzz